requirements.txt
python-telegram-bot[job-queue]==21.6
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-dotenv==1.0.1
Pillow==10.4.0
aiohttp==3.10.10
apscheduler>=3.10
